Low level programming

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,19 +7,19 @@
 int main(void)
 {
 	int i;
-	double f[98];
+	long double f[98];
 
 	f[0] = 1;
 	f[1] = 2;
-	printf("%.0f, %.0f, ", f[0], f[1]);
+	printf("%.0Lf, %.0Lf, ", f[0], f[1]);
 
 	for (i = 2; i < 98; i++)
 	{
 		f[i] = f[i - 1] + f[i - 2];
 		if (i == 97)
-			printf("%.0f\n", f[i]);
+			printf("%.0Lf\n", f[i]);
 		else
-			printf("%.0f, ", f[i]);
+			printf("%.0Lf, ", f[i]);
 	}
 
 	return (0);
